// src/roles/permission.enum.ts
export enum Permission {
  MOVIE_CREATE = 'MOVIE_CREATE',
  MOVIE_UPDATE = 'MOVIE_UPDATE',
  MOVIE_DELETE = 'MOVIE_DELETE',
  VIDEO_UPLOAD = 'VIDEO_UPLOAD',
  USER_MANAGE = 'USER_MANAGE',
  FINANCE_VIEW = 'FINANCE_VIEW',
  DEPOSIT_MANAGE = 'DEPOSIT_MANAGE',
  SUBTITLE_MANAGE = 'SUBTITLE_MANAGE',
  SERIES_MANAGE = 'SERIES_MANAGE',
  SUBSCRIPTION_MANAGE = 'SUBSCRIPTION_MANAGE',
  STAFF_MANAGE = 'STAFF_MANAGE',
}